+++ tests/setup.js
@@ -1,5 +1,0 @@
-module.exports = async () => {
-    process.env.PORTAINER_URL = '';
-    process.env.PORTAINER_USERNAME = '';
-    process.env.PORTAINER_PASSWORD = '';
-};