+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-01-17
+### Added
+- `update-stack-variable` command to update a stack variable.
+
+### Changed
+- Command-line input parsed via `commander` package.
+
 ## [1.3.0] - 2019-01-15
 ### Changed
 - `ALWAYS_UPDATE` swapped for `FORCE_UPDATE`.